restore splash, refactor router selector, move all upgrade stuff to the hook
